Use api of neighbor_add for 4.9 version of Ryu
The api of neighbor_add has been changed in version 4.9 of Ryu.
The previous version sets a capability for address family from result of
parsing address parameter in Ryu.
In the new api, the capability for ipv6 needs to be explicitly specified
by enable_ipv6 parameter.
This patch adds to parse an address in ryu driver in dragent.

Change-Id: I9eafc04a0a4ee651de0376f6cc98957a8a58eaf4